+++ app/src/main/java/com/alessandrosperotti/expandablecardviewexample/MainActivity.kt
@@ -9,7 +9,7 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val card : ExpandableCardView = findViewById(R.id.mycard)
+        val card : com.alespero.expandablecardview.ExpandableCardView = findViewById(R.id.mycard)
 
         card.setOnClickListener({
             if(card.isExpanded) card.collapse()
