+++ app/src/main/res/layout/activity_main.xml
@@ -4,14 +4,15 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="20dp"
     tools:context="com.alessandrosperotti.expandablecardviewexample.MainActivity"
     android:background="#FAFAFA">
 
-<com.alessandrosperotti.expandablecardviewexample.ExpandableCardView
+<com.alespero.expandablecardview.ExpandableCardView
     android:id="@+id/mycard"
+    android:layout_margin="20dp"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
+    app:cardElevation="4dp"
     app:title="Item"
     app:inner_view="@layout/mycustomview"/>
 </LinearLayout>
