+++ expandablecardview/src/main/java/com/alespero/expandablecardview/ExpandableCardView.java
@@ -1,4 +1,4 @@
-package com.alessandrosperotti.expandablecardviewexample;
+package com.alespero.expandablecardview;
 
 
 import android.content.Context;
