+++ expandablecardview/src/main/res/drawable/arrow_down.xml
