+++ expandablecardview/src/main/res/layout/expandable_cardview.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v7.widget.CardView android:id="@+id/card"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    app:cardElevation="4dp"
     android:paddingTop="@dimen/default_vertical_padding"
     android:paddingBottom="@dimen/default_vertical_padding"
     android:descendantFocusability="blocksDescendants"
-    app:cardElevation="4dp"
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
         <LinearLayout
